atention with types loong is needed to initalize the baudrate ant the clock freqeunzie in uart_hal.c and wire_hal.c

+++ includs/uart_hal.h
@@ -4,7 +4,7 @@
 #include "interrupt.h"
 
 
-void uart_init(unsigned int baud, unsigned char fast_mode);
+void uart_init(unsigned long baud, unsigned char fast_mode);
 void uart_send_byte(unsigned char data);
 void uart_send_string(char *data);
 void uart_sendU16(unsigned short val);
